+++ src/test/java/baseball/util/GameNumberGeneratorTest.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.util;
 
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
